added vector of target specification, - only one thing not working: py_list of target specs is not convertible, but we can use .append instead
